edit: some refactoring directorys
