+++ src/modules/role/roletype.enum.ts
@@ -1,5 +1,5 @@
 export enum RoleType {
-   ADMIN = "ADMIN",
+   ADMINISTRATOR = "ADMINISTRATOR",
    AUTHOR = "AUTHOR",
    GENERAL = "GENERAL"
 }
