+++ src/modules/user/test/user.service.spec.ts
@@ -1,15 +1,15 @@
 import { Test, TestingModule } from '@nestjs/testing';
-import { AuthService } from './auth.service';
+import { UserService } from '../user.service';
 
-describe('AuthService', () => {
-  let service: AuthService;
+describe('UserService', () => {
+  let service: UserService;
 
   beforeEach(async () => {
     const module: TestingModule = await Test.createTestingModule({
-      providers: [AuthService],
+      providers: [UserService],
     }).compile();
 
-    service = module.get<AuthService>(AuthService);
+    service = module.get<UserService>(UserService);
   });
 
   it('should be defined', () => {
