+++ src/modules/user/user.controller.ts
@@ -4,6 +4,7 @@ import { User } from './user.entity';
 import { AuthGuard } from '@nestjs/passport';
 import { Roles } from '../role/decorators/role.decorator';
 import { RoleGuard } from '../role/guards/role.guard';
+import { RoleType } from '../role/roletype.enum';
 
 @Controller('users')
 export class UserController {
@@ -12,8 +13,8 @@ export class UserController {
 
    @Get(':id')
    //solo se van a permitir esos roles
-   @Roles('ADMIN', 'GENERAL')
-   @UseGuards(AuthGuard(),RoleGuard)
+   //@Roles(RoleType.ADMINISTRATOR)
+   //@UseGuards(AuthGuard(),RoleGuard)
    async getUser(@Param('id', ParseIntPipe) id: number): Promise<User> {
       const user = await this._userService.get(id);
       return user;
