+++ src/modules/user/user.module.ts
@@ -2,14 +2,13 @@ import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { UserRepository } from './user.repository';
 import { UserService } from './user.service';
-import { SharedModule } from '../../shared/shared.module';
 import { UserController } from './user.controller';
 import { AuthModule } from '../auth/auth.module';
 import { RoleRepository } from '../role/role.repository';
 
 @Module({
    //al importar el sharedModule puedo hacer inyeccion de dependencias del mapper servide en mi modulko usuario
-   imports:[TypeOrmModule.forFeature([UserRepository,RoleRepository]),SharedModule,AuthModule],
+   imports:[TypeOrmModule.forFeature([UserRepository,RoleRepository]),AuthModule],
    providers: [UserService],
    controllers:[UserController]
 })
