--- src/shared/shared.module.ts
@@ -1,8 +1,0 @@
-import { Module } from '@nestjs/common';
-@Module({
-   providers: [],
-   exports: [],
-})
-export class SharedModule {
- 
-}
